1차 완성
git-svn-id: https://dev.naver.com/svn/moonndroid@48 12b37519-7715-0510-82f6-ba5b270cff09

+++ WeekWork/index_db.php
@@ -65,7 +65,7 @@ if ($_REQUEST['select'] == "reply") {
     $results = array();
     while ($array = mysql_fetch_array($result)) {
         $datetime = new DateTime($array['time']);
-        $time = $datetime -> format('y').".".$datetime -> format('m').".".$datetime -> format('d').".".$datetime -> format('H').":".$datetime -> format('i');
+        $time = $datetime -> format('y') . "." . $datetime -> format('m') . "." . $datetime -> format('d') . "." . $datetime -> format('H') . ":" . $datetime -> format('i');
         $results[] = array('reply_id' => $array['reply_id'], 'content' => $array['content'], 'time' => $time, 'user_id' => $array['user_id'], 'work_id' => $array['work_id'], 'file_name' => $array['file_name'], 'file_hash' => $array['file_hash'], 'user_name' => $array['user_name']);
     }
 
@@ -89,32 +89,35 @@ if ($_REQUEST['select'] == "file") {
 }
 
 if ($_REQUEST['select'] == "complete") {
-    mysql_query("UPDATE work SET complete=${_REQUEST['complete']} WHERE work_id=${_REQUEST['work_id']}" );
+    mysql_query("UPDATE work SET complete=${_REQUEST['complete']} WHERE work_id=${_REQUEST['work_id']}");
 }
 
 if ($_REQUEST['select'] == "delWork") {
-    mysql_query("DELETE FROM reply WHERE work_id=${_REQUEST['work_id']}" );
-    mysql_query("DELETE FROM work WHERE work_id=${_REQUEST['work_id']}" );
-    mysql_query("DELETE FROM files WHERE file_id=${_REQUEST['file_id']}" );
-}
-if ($_REQUEST['select'] == "delReply") {
-    mysql_query("DELETE FROM reply WHERE reply_id=${_REQUEST['reply_id']}" );
+    $dir = "./files/";
+    $filehash = $_REQUEST['hash'];
+
+    if (!unlink($dir . $filehash)) {
+        echo "file delete error";
+        exit ;
+    }
+    mysql_query("DELETE FROM work WHERE work_id=${_REQUEST['work_id']}");
+    mysql_query("DELETE FROM files WHERE file_id=${_REQUEST['file_id']}");
 }
 
 if ($_REQUEST['select'] == "sendWork") {
-	if($_REQUEST['work_id']==0){
-	   $sql="INSERT INTO work (work_name,work_content,day,ch_id,user_id) VALUE ('${_REQUEST['work_name']}','${_REQUEST['work_content']}','${_REQUEST['day']}','${_REQUEST['ch_id']}','${_SESSION['id']}')";
-    }else{	    
-	   $sql="UPDATE work SET work_name='${_REQUEST['work_name']}',work_content='${_REQUEST['work_content']}' WHERE work_id='${_REQUEST['work_id']}'";
-	}
-	mysql_query($sql);
+    if ($_REQUEST['work_id'] == 0) {
+        $sql = "INSERT INTO work (work_name,work_content,day,ch_id,user_id) VALUE ('${_REQUEST['work_name']}','${_REQUEST['work_content']}','${_REQUEST['day']}','${_REQUEST['ch_id']}','${_SESSION['id']}')";
+    } else {
+        $sql = "UPDATE work SET work_name='${_REQUEST['work_name']}',work_content='${_REQUEST['work_content']}' WHERE work_id='${_REQUEST['work_id']}'";
+    }
+    mysql_query($sql);
 }
 
 if ($_REQUEST['select'] == "sendReply") {
     $dt = new DateTime();
-    $time= $dt->format('Y-m-d H:i:s');
-	$sql="INSERT INTO reply (content,time,user_id,work_id) VALUE ('${_REQUEST['content']}','${time}','${_SESSION['id']}','${_REQUEST['work_id']}')";
-	mysql_query($sql);
+    $time = $dt -> format('Y-m-d H:i:s');
+    $sql = "INSERT INTO reply (content,time,user_id,work_id) VALUE ('${_REQUEST['content']}','${time}','${_SESSION['id']}','${_REQUEST['work_id']}')";
+    mysql_query($sql);
 }
 ///////////////////////////////////
